tests/ports/psoc-edge/../pdm_pcm: Added simple API function calls test.
